+++ changelog.md
@@ -1,2 +1,9 @@
 ### Fixes
 * Fix innate abilities of imbued trinkets not stacking properly if 2+ trinkets use the same innate ability. (Items crafted before this update will still have this problem unfortunately)
+* Fix fortify missing maximum status effect cap
+
+### Additions
+* Added a reroll button to the enchanting mode of the imbuing table. It is modeled after Easy Magic's reroll mechanic. Enabled by default in configs.
+
+### Changes
+* Bulwark is now an augment, not a standard enchantment. An imbuing recipe for it has been added.